[#3148] Move Proton-j related methods to service-base module
Moved RequestResponseApiConstants' Proton-j related methods to
AbstractRequestResponseEndpoint.
